Added CMOVcc support
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@18 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/core/org/jnode/vm/CpuID.java
@@ -27,5 +27,5 @@ public abstract class CpuID {
 	 * @param feature A platform specific feature constant.
 	 * @return boolean
 	 */
-	public abstract boolean hasFeature(int feature);
+	public abstract boolean hasFeature(long feature);
 }
